chore: trim GUI Ⅱ
