--- GUI/components/__init__.py
@@ -1,1 +1,0 @@
-from GUI.components.menu import ToolMenu
