+++ README.md
@@ -59,7 +59,8 @@
 &emsp;🔳 读剪贴板功能开发中  
 &emsp;🔳 优化速度，翻页等  
 ✅ 增设 rV 按钮（很显眼）及其管理窗口，选脚本然后取消有指示可快捷部署 rV，`显示记录`和`整合章节`移至其窗口内  
-✅ ✨ 将配置系记录系的文件转移位置，后续更新等都不用再手动备份（ 参考[📒配置系文件路径](https://jasoneri.github.io/ComicGUISpider/faq/extra) ）
+✅ ✨ 将配置系记录系的文件转移位置，后续更新等都不用再手动备份（ 参考[📒配置系文件路径](https://jasoneri.github.io/ComicGUISpider/faq/extra) ）  
+✅ ✨ 增加工具视窗，内有`jm`两简化设置域名的标签，受影响的有`hotomi-tools`,`读剪贴板` （ 参考[🎸常规功能](https://jasoneri.github.io/ComicGUISpider/feature)里的图示）  
 
 #### 🐞 Fix
